moved the nav bar components to its own folder
