+++ src/components/navbar/Navbar.jsx
@@ -1,8 +1,16 @@
+
+/**
+ * Navbar component.
+ * 
+ * This component represents the navigation bar of the website.
+ * It includes a logo, a title, and a hamburger menu for mobile devices.
+ * 
+ */
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import NavLinkItem from './NavLinkItem';
+import NavbarItems from './NavbarItems';
 import { FaBars, FaTimes } from 'react-icons/fa';
-import '../cssStyles/Navbar.css'; // Import the CSS file here
+import '../../cssStyles/Navbar.css'; // Import the CSS file here
 
 function Navbar() {
   const [isHamburgerNavOpen, setIsHamburgerNavOpen] = useState(false);
@@ -14,7 +22,7 @@ function Navbar() {
       setAnimateMenu(true);
     } else {
       setAnimateMenu(false);
-      setTimeout(() => setIsHamburgerNavOpen(false), 300); // Match this duration to your CSS animation duration
+      setTimeout(() => setIsHamburgerNavOpen(false), 150); // slight delay to allow animation to finish to better match fade out of menu
     }
   };
 
@@ -26,20 +34,9 @@ function Navbar() {
           <h1 className="text-xl">Glass Gecko Games</h1>
         </Link>
 
-        <ul className="hidden space-x-4 text-primary md:flex">
-          <li>
-            <NavLinkItem to="/">Home</NavLinkItem>
-          </li>
-          <li>
-            <NavLinkItem to="/about">About Us</NavLinkItem>
-          </li>
-          <li>
-            <NavLinkItem to="/news">News</NavLinkItem>
-          </li>
-          <li>
-            <NavLinkItem to="/contact">Contact Us</NavLinkItem>
-          </li>
-        </ul>
+        <div className="hidden md:flex">
+          <NavbarItems textSize="text-lg" />
+        </div>
 
         <button onClick={toggleMenu} className="text-primary block md:hidden z-20">
           {isHamburgerNavOpen ? <FaTimes size={24} /> : <FaBars size={24} />}
@@ -47,20 +44,7 @@ function Navbar() {
 
         {isHamburgerNavOpen && (
           <div className={`fixed md:hidden top-0 left-0 z-10 w-full h-full bg-gray-900 bg-opacity-50 flex justify-center items-center transition-opacity duration-300 ${animateMenu ? 'opacity-100 fade-in' : 'opacity-0 fade-out'}`}>
-            <ul className="text-white text-2xl space-y-4 text-center">
-              <li>
-                <NavLinkItem to="/" onClick={toggleMenu}>Home</NavLinkItem>
-              </li>
-              <li>
-                <NavLinkItem to="/about" onClick={toggleMenu}>About Us</NavLinkItem>
-              </li>
-              <li>
-                <NavLinkItem to="/news" onClick={toggleMenu}>News</NavLinkItem>
-              </li>
-              <li>
-                <NavLinkItem to="/contact" onClick={toggleMenu}>Contact Us</NavLinkItem>
-              </li>
-            </ul>
+            <NavbarItems toggleMenu={toggleMenu} textSize="text-2xl" />
           </div>
         )}
       </div>
